fic database issues

+++ backend_node/src/models/Exam.js
@@ -2,8 +2,8 @@ const { DataTypes } = require('sequelize');
 const sequelize = require('../config/db');
 
 const Exam = sequelize.define('Exam', {
-  id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  school_id: { type: DataTypes.INTEGER, allowNull: false },
+  id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
+  school_id: { type: DataTypes.BIGINT, allowNull: false },
   term_id: { type: DataTypes.INTEGER },
   name: { type: DataTypes.STRING, allowNull: false },
   date: { type: DataTypes.DATE },
