+++ backend_node/src/models/Notification.js
@@ -2,8 +2,8 @@ const { DataTypes } = require('sequelize');
 const sequelize = require('../config/db');
 
 const Notification = sequelize.define('Notification', {
-  id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  school_id: { type: DataTypes.INTEGER, allowNull: false },
+  id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
+  school_id: { type: DataTypes.BIGINT, allowNull: false },
   title: { type: DataTypes.STRING, allowNull: false },
   message: { type: DataTypes.TEXT },
   type: { type: DataTypes.STRING }, // announcement, alert, reminder
